seperated out the functions into files and put in folder
