+++ client/src/components/QnAWidget/Answer.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+const Answer = function(prop) {
+
+  return(
+    <div>
+      <div>A: answer here</div>
+      <div>by NickName at Date</div>
+      <button>helpfull</button>
+      <button>report</button>
+    </div>
+  )
+}
+
+export default Answer;
