+++ client/src/components/QnAWidget/AnswerList.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Answer from './Answer.jsx';
+
+const AnswerList = function(props){
+  const [answerList, setAnswerList] = React.useState([]);
+  React.useEffect(()=>{
+    //test props.alist for length if > 2 setstate to just the first 2 items
+  }, [])
+  return(
+    <div>
+      {answerList.map((item) =>(
+        <Answer answer={item}/>
+      ))}
+      <button>if more answers button</button>
+    </div>
+  )
+}
+export default AnswerList;
