+++ client/src/components/QnAWidget/BottomBar.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const BottomBar = function(props){
+  return(
+    <div>
+      <button>add question</button>
+    </div>
+  )
+}
+export default BottomBar;
