+++ client/src/components/QnAWidget/TopBar.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+const TopBar = function(props){
+
+  return (
+   <div>
+     searchbar here
+
+   </div>
+  )
+ }
+ export default TopBar;
